Item B-03526: As a Manager, I wold like to sort services in the service management table
The ServiceSpecification is implemented to provide sorting and pagination.

The AbstracSpecification used a default 'lastModified' order by, which does not exist in the ServiceRepo.
To prevent breakage, the the order by assignment was move into its own function, toPredicateDefaultQueryOrderBy().
This allows for the ServiceSpecification, as an exception, to not attempt to order by "lastModified" by overriding that new method.