Rename `coverage.yml` to `build.yml` (and 'Coverage' to 'Build').
The reason being that we still want a Build Badge.
The Coverage Badge is coming from coveralls.
There is a desire to have only a single Workflow.

To achieve this, the build.yml that builds AND sends to coveralls must be used as a Build Badge.
This is because the GithubActions will builds the Badges entirely based on the name provided by the Workflow.
The Workflow file itself is also renamed to be consistent with the name property within the Workflow.